requirements.txt
chardet==4.0.0
et-xmlfile==1.1.0
lml==0.1.0
openpyxl==3.0.9
pyexcel==0.6.7
pyexcel-io==0.6.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
rsl==0.2.1
texttable==1.6.4
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.4.0

python-dotenv~=0.19.1
terminaltables~=3.1.0